requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-anthropic>=0.1.0
anthropic>=0.25.0
tavily-python>=0.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0
httpx>=0.27.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.9.0
orjson>=3.9.0
loguru>=0.7.0
tiktoken>=0.7.0
tenacity>=8.2.0
fastapi>=0.110.0
uvicorn>=0.29.0
ipykernel>=6.29.0
rich>=13.7.0